create camera c file
